split layout.tsx to multiple smaller components
